Split generate_music_decode tests to satisfy 200 LOC cap.
Move shared fixtures into generate_music_decode_test_support.py and split
scenario tests into prepare, basic, CPU-offload, and VRAM modules.
